fix : 문서 - 마지막 편집 시간을 간단하게 표기하도록 수정
기존에는 n일 n시간 n분 전 까지 표시되었지만 지금은 n일 전, n시간 전, n분 전, n초 전 으로 나누어 표시
